requirements.txt
Automat>=0.8.0
Click>=7.0
Django>=3.0.2
Flask>=1.1.1
Jinja2>=2.10.3
MarkupSafe>=1.1.1
Protego>=0.1.16
PyDispatcher>=2.0.5
PyHamcrest>=1.9.0,<2.0.0
PyMySQL>=0.9.3
Routes>=2.4.1
SQLAlchemy>=1.3.12
Scrapy>=1.8.0
Twisted>=19.10.0
Werkzeug>=0.16.0
argparse>=1.4.0
asgiref>=3.2.3
attrs>=19.3.0
automium>=0.2.6
automium-web>=0.1.1
beautifulsoup4>=4.8.2
bs4>=0.0.1
certifi>=2019.11.28
cffi>=1.13.2
chardet>=3.0.4
constantly>=15.1.0
coreapi>=2.3.3
coreschema>=0.0.4
cryptography>=2.8
cssselect>=1.1.0
cycler>=0.10.0
demjson>=2.2.4
django-cors-headers>=3.2.1
django-rest-framework>=0.1.0
django-tastypie>=0.14.3
djangorestframework>=3.11.0
hyperlink>=19.0.0
idna>=2.8
incremental>=17.5.0
itsdangerous>=1.1.0
itypes>=1.1.0
kiwisolver>=1.1.0
legacy>=0.1.6
lxml>=4.4.2
matplotlib>=3.1.2
mysqlclient>=1.4.6
nose>=1.3.7
numpy>=1.18.1
pandas>=0.25.3
parsel>=1.5.2
pip>=19.0.3,<19.3.1
post>=2019.4.13
psycopg2>=2.8.4
pyOpenSSL>=19.1.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.19
pyparsing>=2.4.6
python-dateutil>=2.8.1
python-mimeparse>=1.6.0
pytz>=2019.3
queuelib>=1.5.0
repoze.lru>=0.7
requests>=2.22.0
scipy>=1.4.1
seaborn>=0.9.0
service-identity>=18.1.0
setuptools>=40.8.0,<45.0.0
six>=1.13.0,<1.14.0
soupsieve>=1.9.5
sqlparse>=0.3.0
toolz>=0.10.0
uritemplate>=3.0.1
urllib3>=1.25.7
views>=0.3
w3lib>=1.21.0
xlrd>=1.2.0
xmltodict>=0.12.0
zope.interface>=4.7.1